LayoutTests/platform/chromium-cg-mac/fast/events/mouseup-from-button2-expected.txt

This test verifies that a right mousebutton click generates a mousedown, context and mouseup event.

If the test passes, you'll see a PASS message below.

To run this test in Safari:

Right mouse down inside the box.
Right mouse up.
box

type: mousedown, button: 0, 
type: mouseup, button: 0, 
type: click, button: 0, 
type: mousedown, button: 2, 
type: contextmenu, button: 2, 
type: mouseup, button: 2, 
type: mousedown, button: 2, 
type: contextmenu, button: 2, 
